+++ src/routes/venue/[venueid]/+layout.svelte
@@ -2,9 +2,9 @@
 	import type { Snippet } from 'svelte';
 	import type { LayoutData } from './$types';
 	import Page from '$lib/components/Page.svelte';
-	import Note from '$lib/components/Note.svelte';
 	import Link from '$lib/components/Link.svelte';
 	import ScholarLink from '$lib/components/ScholarLink.svelte';
+	import Feedback from '$lib/components/Feedback.svelte';
 
 	let { data, children }: { data: LayoutData; children: Snippet } = $props();
 	const { venue } = $derived(data);
@@ -20,15 +20,14 @@
 		{#snippet details()}<Link to={venue.url}>{venue.url}</Link> Editors: {#each venue.editors as editorID}
 				<ScholarLink id={editorID} />
 			{/each}{/snippet}
-		<Note error><strong>{venue.inactive}</strong></Note>
+		<Feedback error><strong>{venue.inactive}</strong></Feedback>
 	</Page>
 {:else}
 	{#if venue.inactive !== null}
-		<Note error
-			>This venue is currently inactive. Only you and other editors can see it, whereas everyone
-			else sees this message: "{venue.inactive}". When you're ready, activate it in <Link
-				to="/venue/{venue.id}/settings">venue settings</Link
-			>.</Note
+		<Feedback error
+			>This venue is currently <strong>inactive</strong>. Only you and other editors can see it,
+			along with the message set in the <Link to="/venue/{venue.id}/settings">venue settings</Link>.
+			Visit the settings to activate the venue.</Feedback
 		>
 	{/if}
 	{@render children()}