+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Dates should be in`YYYY-MM-DD` format.
 - Added venue active/inactive switch, to allow for venues to be created without being availabe for use yet.
 - Created editor-guarded settings page to organize venue settings and declutter landing page.
 
+### Fixed
+
+- Eliminated redundant error `Note`s, merging them with `Feedback`.
+
 ## 0.2.2 2026-01-18
 
 ### Added